requirements.txt
asgiref==3.6.0
aspose-pdf==23.2.0
Babel==2.12.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
cryptography==39.0.2
cssselect==1.2.0
Django==4.1.6
django-countries==7.5.1
djangorestframework==3.14.0
docxcompose==1.4.0
fpdf==1.7.2
idna==3.4
lxml==4.9.2
pdfminer.six==20221105
pdfquery==0.4.3
pycparser==2.21
pypdf2==3.0.1
pyquery==2.0.0
python-docx==0.8.11
python-docx-replace==0.4.2
pytz==2022.7.1
requests==2.28.2
roman==4.0
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.5.0
urllib3==1.26.14
